+++ src/sentry/android/android_string_names.h
@@ -37,7 +37,6 @@ public:
 	StringName removeUser;
 	StringName addBreadcrumb;
 	StringName log;
-	StringName captureMessage;
 	StringName getLastEventId;
 	StringName captureError;
 	StringName createEvent;