+++ src/sentry/javascript/javascript_sdk.cpp
@@ -1,5 +1,6 @@
 #include "javascript_sdk.h"
 
+#include "sentry/disabled/disabled_scope.h"
 #include "sentry/javascript/javascript_breadcrumb.h"
 #include "sentry/javascript/javascript_event.h"
 #include "sentry/javascript/javascript_interop.h"
@@ -164,7 +165,7 @@ void JavaScriptSDK::add_breadcrumb(const Ref<SentryBreadcrumb> &p_breadcrumb) {
 	js_bridge()->call("addBreadcrumb", crumb->get_js_object());
 }
 
-void JavaScriptSDK::log(LogLevel p_level, const String &p_body, const Dictionary &p_attributes) {
+void JavaScriptSDK::capture_log(const Ref<SentryScope> &p_scope, LogLevel p_level, const String &p_body, const Dictionary &p_attributes) {
 	ERR_FAIL_COND(!js_bridge());
 
 	String attr_value = attributes_to_json(p_attributes);
@@ -191,12 +192,6 @@ void JavaScriptSDK::log(LogLevel p_level, const String &p_body, const Dictionary
 	}
 }
 
-String JavaScriptSDK::capture_message(const String &p_message, Level p_level) {
-	ERR_FAIL_COND_V(!js_bridge(), String());
-	String uuid = js_bridge()->call("captureMessage", p_message.utf8(), level_as_string(p_level).utf8()).as_string();
-	return uuid;
-}
-
 String JavaScriptSDK::get_last_event_id() {
 	ERR_FAIL_COND_V(!js_bridge(), String());
 	return js_bridge()->call("lastEventId").as_string();
@@ -206,7 +201,7 @@ Ref<SentryEvent> JavaScriptSDK::create_event() {
 	return memnew(JavaScriptEvent);
 }
 
-String JavaScriptSDK::capture_event(const Ref<SentryEvent> &p_event) {
+String JavaScriptSDK::capture_event(const Ref<SentryScope> &p_scope, const Ref<SentryEvent> &p_event) {
 	ERR_FAIL_COND_V(!js_bridge(), String());
 	ERR_FAIL_COND_V_MSG(p_event.is_null(), String(), "Sentry: Can't capture event - event object is null.");
 	JavaScriptEvent *ev = Object::cast_to<JavaScriptEvent>(p_event.ptr());
@@ -214,7 +209,7 @@ String JavaScriptSDK::capture_event(const Ref<SentryEvent> &p_event) {
 	return js_bridge()->call("captureEvent", ev->get_js_object()).as_string();
 }
 
-void JavaScriptSDK::capture_feedback(const Ref<SentryFeedback> &p_feedback) {
+void JavaScriptSDK::capture_feedback(const Ref<SentryScope> &p_scope, const Ref<SentryFeedback> &p_feedback) {
 	ERR_FAIL_COND(!js_bridge());
 	ERR_FAIL_COND_MSG(p_feedback.is_null(), "Sentry: Can't capture feedback - feedback object is null.");
 	ERR_FAIL_COND_MSG(p_feedback->get_message().is_empty(), "Sentry: Can't capture feedback - feedback message is empty.");
@@ -252,19 +247,19 @@ void JavaScriptSDK::clear_attachments() {
 	js_bridge()->call("clearAttachments");
 }
 
-void JavaScriptSDK::metrics_add_count(const String &p_name, int64_t p_value, const Dictionary &p_attributes) {
+void JavaScriptSDK::metrics_add_count(const Ref<SentryScope> &p_scope, const String &p_name, int64_t p_value, const Dictionary &p_attributes) {
 	ERR_FAIL_COND(!js_bridge());
 	String attr_value = attributes_to_json(p_attributes);
 	js_bridge()->call("metricsAddCount", p_name.utf8(), p_value, attr_value.utf8());
 }
 
-void JavaScriptSDK::metrics_add_gauge(const String &p_name, double p_value, const String &p_unit, const Dictionary &p_attributes) {
+void JavaScriptSDK::metrics_add_gauge(const Ref<SentryScope> &p_scope, const String &p_name, double p_value, const String &p_unit, const Dictionary &p_attributes) {
 	ERR_FAIL_COND(!js_bridge());
 	String attr_value = attributes_to_json(p_attributes);
 	js_bridge()->call("metricsAddGauge", p_name.utf8(), p_value, p_unit.utf8(), attr_value.utf8());
 }
 
-void JavaScriptSDK::metrics_add_distribution(const String &p_name, double p_value, const String &p_unit, const Dictionary &p_attributes) {
+void JavaScriptSDK::metrics_add_distribution(const Ref<SentryScope> &p_scope, const String &p_name, double p_value, const String &p_unit, const Dictionary &p_attributes) {
 	ERR_FAIL_COND(!js_bridge());
 	String attr_value = attributes_to_json(p_attributes);
 	js_bridge()->call("metricsAddDistribution", p_name.utf8(), p_value, p_unit.utf8(), attr_value.utf8());
@@ -294,6 +289,10 @@ void JavaScriptSDK::set_attribute(const String &p_name, const Variant &p_value) 
 void JavaScriptSDK::remove_attribute(const String &p_name) {
 	ERR_FAIL_COND(!js_bridge());
 	js_bridge()->call("removeAttribute", p_name.utf8());
+}
+
+SentryScopeImpl *JavaScriptSDK::create_scope() {
+	return memnew(DisabledScope);
 }
 
 void JavaScriptSDK::set_trace(const String &p_trace_id, const String &p_parent_span_id) {